Created and implimented user and invalid user exception classes.
User -
This class will contain all logic for the users, such as capitalising names and validating the email.

InvalidUserException.php -
This is used when an error is thrown due to the email being invalid, it can be extended for other use cases in the future.

Updated the readme to include a project structure and future improvements.

Updated user_upload.php to use the new classes.